CARDS-2910: Move specialized label processors into their question-type modules
The base label processors are moved to the exported package
io.uhndata.cards.forms.serialize.labels so other bundles can extend them. The
vocabulary, pedigree and DICOM label processors then move out of data-model into
their own modules, where each is loaded only in test mode together with its
question type. The core label processors stay in data-model.
