+++ modules/dicom/src/main/java/io/uhndata/cards/dicom/internal/DicomLabelProcessor.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.uhndata.cards.forms.internal.serialize.labels;
+package io.uhndata.cards.dicom.internal;
 
 import java.util.function.Function;
 
@@ -28,6 +28,7 @@ import jakarta.json.JsonValue;
 
 import org.osgi.service.component.annotations.Component;
 
+import io.uhndata.cards.forms.serialize.labels.FileLabelProcessor;
 import io.uhndata.cards.serialize.spi.ResourceJsonProcessor;
 
 /**
