+++ modules/pedigree/src/main/java/io/uhndata/cards/pedigree/internal/PedigreeLabelProcessor.java
@@ -1,0 +1,71 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package io.uhndata.cards.pedigree.internal;
+
+import java.util.function.Function;
+
+import javax.jcr.Node;
+import javax.jcr.RepositoryException;
+
+import jakarta.json.Json;
+import jakarta.json.JsonObjectBuilder;
+import jakarta.json.JsonValue;
+
+import org.osgi.service.component.annotations.Component;
+
+import io.uhndata.cards.forms.serialize.labels.SimpleAnswerLabelProcessor;
+import io.uhndata.cards.serialize.spi.ResourceJsonProcessor;
+
+/**
+ * Gets the pedigree question answer as svg picture.
+ *
+ * @version $Id$
+ */
+@Component(immediate = true)
+public class PedigreeLabelProcessor extends SimpleAnswerLabelProcessor implements ResourceJsonProcessor
+{
+    @Override
+    public String getDescription()
+    {
+        return "Get the answers for pedigree questions as an svg picture.";
+    }
+
+    @Override
+    public void leave(Node node, JsonObjectBuilder json, Function<Node, JsonValue> serializeNode)
+    {
+        try {
+            if (node.isNodeType("cards:PedigreeAnswer")) {
+                addProperty(node, json, serializeNode);
+            }
+        } catch (RepositoryException e) {
+            // Really shouldn't happen
+        }
+    }
+
+    @Override
+    public JsonValue getAnswerLabel(final Node node, final Node question)
+    {
+        try {
+            return Json.createValue(node.getProperty("image").getValue().toString());
+        } catch (RepositoryException e) {
+            // Really shouldn't happen
+        }
+        return null;
+    }
+}
