+++ modules/vocabularies/src/main/java/io/uhndata/cards/vocabularies/internal/VocabularyOptionsLabelProcessor.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.uhndata.cards.forms.internal.serialize.labels;
+package io.uhndata.cards.vocabularies.internal;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -33,6 +33,7 @@ import jakarta.json.JsonValue;
 import org.apache.sling.api.resource.Resource;
 import org.osgi.service.component.annotations.Component;
 
+import io.uhndata.cards.forms.serialize.labels.SimpleAnswerLabelProcessor;
 import io.uhndata.cards.serialize.spi.ResourceJsonProcessor;
 
 /**
